+++ flower/__init__.py
@@ -1,2 +1,2 @@
-VERSION = (1, 0, 1)
+VERSION = (2, 0, 0)
 __version__ = '.'.join(map(str, VERSION)) + '-dev'
