+++ flower/templates/navbar.html
@@ -11,7 +11,7 @@
   <div class="collapse navbar-collapse" id="navbarSupportedContent">
     <ul class="navbar-nav mr-auto">
       <li class="nav-item active">
-        <a class="nav-link text-dark" href="{{ reverse_url('dashboard') }}">Workers</a>
+        <a class="nav-link text-dark" href="{{ reverse_url('workers') }}">Workers</a>
       </li>
       <li class="nav-item">
         <a class="nav-link text-dark" href="{{ reverse_url('tasks') }}">Tasks</a>
