+++ flower/urls.py
@@ -11,7 +11,7 @@ from .views.broker import BrokerView
 from .views.workers import WorkerView
 from .views.tasks import TaskView, TasksView, TasksDataTable
 from .views.error import NotFoundErrorHandler
-from .views.dashboard import DashboardView
+from .views.workers import WorkersView
 from .utils import gen_cookie_secret
 
 
@@ -26,8 +26,8 @@ settings = dict(
 
 handlers = [
     # App
-    url(r"/", DashboardView, name='main'),
-    url(r"/dashboard", DashboardView, name='dashboard'),
+    url(r"/", WorkersView, name='main'),
+    url(r"/workers", WorkersView, name='workers'),
     url(r"/worker/(.+)", WorkerView, name='worker'),
     url(r"/task/(.+)", TaskView, name='task'),
     url(r"/tasks", TasksView, name='tasks'),
